BACKEND/requirements.txt
fastapi
uvicorn
python-multipart
pandas
requests
openpyxl
mangaba_ai
python-dotenv
google-generativeai
tabulate
